refactor(*): remove unused code
