+++ apps/webaudio-spectrum-analyser/src/app/components/audio-controls/audio-controls.component.spec.ts
@@ -5,7 +5,7 @@ import { ClarityModule } from '@clr/angular';
 
 import { SafeUrlPipe } from '../../pipes/safe-url/safe-url.pipe';
 import { TimePipe } from '../../pipes/time/time.pipe';
-import { mockComponent } from '../../utils/test';
+import { mockComponent } from '../../utils/test.util';
 import { AlertComponent } from '../alert/alert.component';
 import { AudioControlsComponent } from './audio-controls.component';
