+++ apps/webaudio-spectrum-analyser/src/app/utils/factories.ts
@@ -20,10 +20,3 @@ export function getDocument() {
 export function getAudioGraph() {
   return new AudioGraph();
 }
-
-/**
- * Audio graph factory
- */
-export function getMockAudioGraph() {
-  return new AudioGraph(true);
-}
