--- tools/shell/build-codepen
@@ -1,46 +1,0 @@
-#!/bin/bash
-
-SRC_DIR='apps/webaudio-spectrum-analyser/src'
-
-DIST_DIR='dist/codepen'
-DIST_HTML="${DIST_DIR}/index.html"
-DIST_CSS="${DIST_DIR}/style.scss"
-DIST_TS="${DIST_DIR}/script.ts"
-
-cd "$(dirname "${BASH_SOURCE[0]}")/../.."
-source tools/shell/common.sh
-
-[[ -e $DIST_DIR ]] && rm -rfv "${DIST_DIR}/*"
-mkdir -pv "$DIST_DIR" || error 'mkdir failed'
-
-log 'building html...'
-
-exec 3>"$DIST_HTML" || error "could not open '${DIST_HTML}'"
-echo '<app-root></app-root>' >&3 || error "could not write '${DIST_HTML}'"
-while read path; do
-  id="$(basename $path)"
-  echo -e "\\n\\n<script type=\"text/ng-template\" id=\"${id}\">" >&3 \
-    && cat "$path" >&3 \
-    && echo '</script>' >&3 \
-    || error "could not add '${path}' to '${DIST_HTML}'"
-done < <(find "${SRC_DIR}" -iname '*.component.html');
-exec 3>&- || error "could not close '${DIST_HTML}'"
-
-log 'building css...'
-
-egrep \
-  -hv '^\s*@import' \
-  "${SRC_DIR}/css/"{_vars,vendor,styles}".scss" \
-  >"$DIST_CSS" \
-  || error 'css build failed'
-
-log 'building ts...'
-
-find "$SRC_DIR" \
-     -iname '*.ts' \
-     -not -iname '*.spec.ts' \
-     -not -iname 'test-setup.ts' \
-     -not -iname '*.prod.ts' \
-     -not -iname 'polyfills.ts' \
-  | xargs ./tools/shell/ts-concat ./tools/shell/modules.json >"$DIST_TS" \
-  || error 'ts build failed'
