--- tools/shell/modules.json
@@ -1,16 +1,0 @@
-{
-  "@angular/common": "ng.common",
-  "@angular/core": "ng.core",
-  "@angular/forms": "ng.forms",
-  "@angular/platform-browser": "ng.platformBrowser",
-  "@angular/platform-browser-dynamic": "ng.platformBrowserDynamic",
-  "@angular/platform-browser/animations": "ng.platformBrowser.animations",
-  "@angular/router": "ng.router",
-  "@clr/angular": "clr.angular",
-  "simplebar-angular": "window['simplebar-angular']",
-  "@ngxs/store": "ngxs.store",
-  "@ngxs/store/operators": "ngxs.store.operators",
-  "@ngxs/logger-plugin": "ngxs['logger-plugin']",
-  "rxjs": "rxjs",
-  "rxjs/operators": "rxjs.operators"
-}
